+++ vaadin-chip-combobox/src/main/java/software/xdev/vaadin/chips/ChipComboBox.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2021 XDEV Software (https://xdev.software/en)
+ * Copyright © 2021 XDEV Software (https://xdev.software)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,27 +20,30 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import com.vaadin.flow.component.AbstractCompositeField;
 import com.vaadin.flow.component.AbstractField.ComponentValueChangeEvent;
 import com.vaadin.flow.component.ClickEvent;
 import com.vaadin.flow.component.Component;
+import com.vaadin.flow.component.Focusable;
+import com.vaadin.flow.component.HasLabel;
 import com.vaadin.flow.component.HasSize;
 import com.vaadin.flow.component.HasStyle;
+import com.vaadin.flow.component.HasTheme;
 import com.vaadin.flow.component.HasValidation;
 import com.vaadin.flow.component.ItemLabelGenerator;
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.button.ButtonVariant;
 import com.vaadin.flow.component.combobox.ComboBox;
-import com.vaadin.flow.component.combobox.dataview.ComboBoxDataView;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.internal.AbstractFieldSupport;
 import com.vaadin.flow.component.orderedlayout.FlexComponent;
 import com.vaadin.flow.component.orderedlayout.FlexLayout;
+import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.select.Select;
+import com.vaadin.flow.component.shared.HasTooltip;
 import com.vaadin.flow.data.binder.HasItems;
 import com.vaadin.flow.data.provider.DataChangeEvent;
 import com.vaadin.flow.data.provider.DataProvider;
@@ -48,30 +51,34 @@ import com.vaadin.flow.data.provider.DataView;
 import com.vaadin.flow.data.provider.HasDataView;
 import com.vaadin.flow.data.provider.InMemoryDataProvider;
 import com.vaadin.flow.data.provider.Query;
+import com.vaadin.flow.data.binder.HasValidator;
 import com.vaadin.flow.dom.Style;
 import com.vaadin.flow.function.SerializableFunction;
 
 
 /**
- * This component has a combobox with available items and displays the selected items as "chips" underneath it.<br/> It
+ * This component has a ComboBox with available items and displays the selected items as "chips" underneath it.<br/> It
  * behaves somewhat similar to a {@link Select}.
- *
- * @author DL
- * @author JohannesRabauer
- * @author AB
  */
 public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, ChipComboBox<T>, Collection<T>> implements
-	HasItems<T>,
-	HasValidation,
 	HasStyle,
+	Focusable<ChipComboBox<T>>,
 	HasSize,
-	HasDataView<T, String, ComboBoxDataView<T>>
+	HasValidation,
+	HasTheme,
+	HasLabel,
+	HasTooltip,
+	HasItems<T>,,
+	HasDataView<T, String, ComboBoxDataView<T>>,
+	HasValidator<T>
 {
 	
 	/*
 	 * UI-Components
 	 */
 	protected ComboBox<T> cbAvailableItems = new ComboBox<>();
+	protected Button btnClearAll = new Button(VaadinIcon.TRASH.create());
+	protected HorizontalLayout comboBoxContainer = new HorizontalLayout();
 	protected FlexLayout chipsContainer = new FlexLayout();
 	protected Button btnClearAll = new Button(VaadinIcon.TRASH.create());
 	
@@ -102,14 +109,16 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		chipsContainerStyle.set("flex-direction", "row");
 		
 		this.btnClearAll.addThemeVariants(ButtonVariant.LUMO_SMALL, ButtonVariant.LUMO_TERTIARY_INLINE);
-		final FlexLayout comboBoxAndClearButton = new FlexLayout(this.cbAvailableItems, this.btnClearAll);
-		comboBoxAndClearButton.setAlignItems(FlexComponent.Alignment.BASELINE);
+		
+		this.comboBoxContainer.setAlignItems(FlexComponent.Alignment.BASELINE);
+		this.comboBoxContainer.setWidthFull();
+		this.comboBoxContainer.add(this.cbAvailableItems, this.btnClearAll);
 		
 		this.getContent().setPadding(false);
 		this.getContent().setSpacing(false);
 		this.setSizeUndefined();
 		
-		this.getContent().add(comboBoxAndClearButton, this.chipsContainer);
+		this.getContent().add(this.comboBoxContainer, this.chipsContainer);
 		
 		// Since version 2.2 the default
 		// Long words (> 18 chars) are not displayed correctly with the hardcoded width of 12em
@@ -122,16 +131,6 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.btnClearAll.addClickListener(this::onClickClearAll);
 	}
 	
-	protected void onClickClearAll(final ClickEvent<Button> buttonClickEvent)
-	{
-		if(this.isReadOnly())
-		{
-			return;
-		}
-		final List<T> values = new ArrayList<>();
-		this.updateValues(values, buttonClickEvent.isFromClient());
-	}
-	
 	protected void onCbAvailableItemsValueChanged(final ComponentValueChangeEvent<ComboBox<T>, T> event)
 	{
 		if(event.getValue() == null || this.isReadOnly())
@@ -140,6 +139,16 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		}
 		
 		this.addItem(event.getValue(), event.isFromClient());
+	}
+	
+	protected void onClickClearAll(final ClickEvent<Button> event)
+	{
+		if(this.isReadOnly())
+		{
+			return;
+		}
+		
+		this.updateValues(this.getEmptyValue(), event.isFromClient());
 	}
 	
 	@Override
@@ -153,7 +162,6 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.selectedComponents.removeIf(comp -> !newPresentationValue.contains(comp.getItem()));
 		
 		// Find new values and build components
-		// @formatter:off
 		final Collection<T> existingValues =
 			this.selectedComponents.stream()
 				.map(ChipComponent::getItem)
@@ -177,7 +185,6 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 				return chipComponent;
 			})
 			.forEach(this.selectedComponents::add);
-		// @formatter:on
 		
 		this.updateUI();
 	}
@@ -230,6 +237,7 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	{
 		this.updateSelectedChips();
 		this.updateAvailableItems();
+		this.updateRequiredIndicatorOfCbAvailableItems();
 	}
 	
 	protected void updateSelectedChips()
@@ -243,6 +251,11 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		final List<T> availableItems = new ArrayList<>(this.allAvailableItems);
 		availableItems.removeAll(this.getValue());
 		this.cbAvailableItems.setItems(availableItems);
+	}
+	
+	protected void updateRequiredIndicatorOfCbAvailableItems()
+	{
+		this.cbAvailableItems.setRequiredIndicatorVisible(this.isRequiredIndicatorVisible());
 	}
 	
 	/**
@@ -266,13 +279,7 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.updateUI();
 	}
 	
-	///////////////////////////////////////////////////////////////////////////
-	// setters + getters //
-	///////////////////////
-	
-	/*
-	 * Chips Supplier
-	 */
+	// region Chips Supplier
 	
 	/**
 	 * Returns the current supplier for creating new {@link ChipComponent ChipComponents}
@@ -285,10 +292,8 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	}
 	
 	/**
-	 * @see ChipComboBox#setChipsSupplier(Supplier)
-	 *
-	 * @param chipsSupplier
-	 * @return the component itself
+	 * @return self
+	 * @see #setChipsSupplier(SerializableFunction)
 	 */
 	public ChipComboBox<T> withChipsSupplier(final SerializableFunction<T, ChipComponent<T>> chipsSupplier)
 	{
@@ -299,22 +304,18 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	/**
 	 * Sets the supplier for creating new {@link ChipComponent ChipComponents}
 	 *
-	 * @param chipsSupplier
-	 *            supplier for creating new {@link ChipComponent ChipComponents}
+	 * @param chipsSupplier supplier for creating new {@link ChipComponent ChipComponents}
 	 */
 	public void setChipsSupplier(final SerializableFunction<T, ChipComponent<T>> chipsSupplier)
 	{
 		this.chipsSupplier = Objects.requireNonNull(chipsSupplier);
 	}
 	
-	/*
-	 * All available items
-	 */
+	// endregion
+	// region All available items
 	
 	/**
 	 * Get all available items, that can potentially get selected
-	 *
-	 * @return
 	 */
 	public List<T> getAllAvailableItems()
 	{
@@ -328,10 +329,13 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		return this;
 	}
 	
-	/*
-	 * Label (of the ComboBox)
-	 */
+	// endregion
+	// region Label (of the ComboBox)
 	
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
 	public String getLabel()
 	{
 		return this.cbAvailableItems.getLabel();
@@ -343,14 +347,17 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		return this;
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
 	public void setLabel(final String label)
 	{
 		this.cbAvailableItems.setLabel(label);
 	}
 	
-	/*
-	 * Placeholder (of the ComboBox)
-	 */
+	// endregion
+	// region Placeholder (of the ComboBox)
 	
 	public String getPlaceholder()
 	{
@@ -368,9 +375,8 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.cbAvailableItems.setPlaceholder(placeholder);
 	}
 	
-	/*
-	 * En-/Disable ClearAll-Button
-	 */
+	// endregion
+	// region ClearAll-Button
 	
 	/**
 	 * @return "Clear All Button" visibility. With this button it is possible to clear all selected items with one
@@ -382,13 +388,10 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	}
 	
 	/**
-	 * Sets the "Clear All Button" to visible or invisible. With this button it is possible to clear all selected items
-	 * with one click. The default value is {@code true}.
-	 *
-	 * @param clearAllButtonVisible defines the visibility of the "Clear All Button".
 	 * @return self
+	 * @see #setClearAllButtonVisible(boolean)
 	 */
-	public ChipComboBox<T> withIsClearAllButtonVisible(final boolean clearAllButtonVisible)
+	public ChipComboBox<T> withClearAllButtonVisible(final boolean clearAllButtonVisible)
 	{
 		this.setClearAllButtonVisible(clearAllButtonVisible);
 		return this;
@@ -405,10 +408,6 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.btnClearAll.setVisible(clearAllButtonVisible);
 	}
 	
-	/*
-	 * Icon of ClearAll-Button
-	 */
-	
 	/**
 	 * @return "Clear All Button" icon. With this button it is possible to clear all selected items with one click. The
 	 * default value is {@link VaadinIcon#TRASH}.
@@ -419,12 +418,8 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	}
 	
 	/**
-	 * Sets the "Clear All Button" icon. With this button it is possible to clear all selected items with one click.
-	 * The
-	 * default value is {@link VaadinIcon#TRASH}.
-	 *
-	 * @param clearAllIcon the "Clear All Button" icon.
 	 * @return self
+	 * @see #setClearAllIcon(Component)
 	 */
 	public ChipComboBox<T> withClearAllIcon(final Component clearAllIcon)
 	{
@@ -444,9 +439,8 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.btnClearAll.setIcon(clearAllIcon);
 	}
 	
-	/*
-	 * FullComboBoxWidth
-	 */
+	// endregion
+	// region FullComboBoxWidth
 	
 	public ChipComboBox<T> withFullComboBoxWidth(final boolean useFullWidth)
 	{
@@ -466,9 +460,8 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		}
 	}
 	
-	/*
-	 * Item Label generator
-	 */
+	// endregion
+	// region Item Label generator
 	
 	/**
 	 * Sets the item label generator used by the individual {@link ChipComponent}s.
@@ -498,8 +491,6 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	/**
 	 * Convenience method, which sets the item label generator used by *BOTH* {@link ComboBox} and the
 	 * {@link ChipComponent}s.
-	 *
-	 * @return
 	 */
 	public void setItemLabelGenerator(final ItemLabelGenerator<T> generator)
 	{
@@ -520,10 +511,15 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		return this;
 	}
 	
+	// endregion
+	
 	/*
 	 * Other
 	 */
 	
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
 	public void setValue(final Collection<T> value)
 	{
@@ -538,6 +534,9 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		super.setValue(value);
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
 	public void setReadOnly(final boolean readOnly)
 	{
@@ -547,12 +546,25 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		this.selectedComponents.forEach(comp -> comp.setReadonly(readOnly));
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 */
 	@Override
 	public void setRequiredIndicatorVisible(final boolean requiredIndicatorVisible)
 	{
 		super.setRequiredIndicatorVisible(requiredIndicatorVisible);
 		
-		this.cbAvailableItems.setRequiredIndicatorVisible(requiredIndicatorVisible);
+		this.updateRequiredIndicatorOfCbAvailableItems();
+	}
+	
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public boolean isRequiredIndicatorVisible()
+	{
+		// The required-indicator should only be visible when it's defined and empty
+		return super.isRequiredIndicatorVisible() && this.isEmpty();
 	}
 	
 	/**
@@ -591,13 +603,32 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 		return this.cbAvailableItems.isInvalid();
 	}
 	
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public void focus()
+	{
+		this.getCbAvailableItems().focus();
+	}
+	
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public void blur()
+	{
+		this.getCbAvailableItems().blur();
+	}
+	
 	/*
 	 * UI-Components
 	 */
 	
 	/**
-	 * Returns the {@link ComboBox} which contains the available items.<br/>
-	 * NOTE: If the contents of the {@link ComboBox} are modified from the outside this component may break
+	 * Returns the {@link ComboBox} which contains the available items.
+	 * <p/>
+	 * NOTE: If the contents are modified from the outside this component may break
 	 */
 	public ComboBox<T> getCbAvailableItems()
 	{
@@ -605,8 +636,19 @@ public class ChipComboBox<T> extends AbstractCompositeField<VerticalLayout, Chip
 	}
 	
 	/**
-	 * Returns the {@link FlexLayout} with the select items (as {@link ChipComponent}s).<br/> NOTE: If the contents of
-	 * the {@link FlexLayout} are modified from the outside this component may break
+	 * Returns the {@link HorizontalLayout} which contains {@link #cbAvailableItems}.
+	 * <p/>
+	 * NOTE: If the contents are modified from the outside this component may break
+	 */
+	public HorizontalLayout getComboBoxContainer()
+	{
+		return this.comboBoxContainer;
+	}
+	
+	/**
+	 * Returns the {@link FlexLayout} with the select items (as {@link ChipComponent}s).
+	 * <p/>
+	 * NOTE: If the contents are modified from the outside this component may break
 	 */
 	public FlexLayout getChipsContainer()
 	{