+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.1.0
+
+* Added a clear-all button
+* Implemented more Vaadin-Mixins for better customization
+* Fixed the required indicator
+  * The indicator of the "available items"-ComboBox is now only present when the wrapping ChipComboBox is empty
+  * The indicator value is now correctly returned by ``isRequiredIndicatorVisible``
+* Updated dependencies
+
 ## 3.0.1
 
 * Updated dependencies